requirements.txt
requests
python-dotenv
streamlit
pandas
plotly
statsmodels
scikit-learn
